+++ CHANGES.rst
@@ -5,4 +5,6 @@ Changelog of lizard-svg
 0.1 (unreleased)
 ----------------
 
+- Added dummy timeseries api.
+
 - Initial library skeleton created by nensskel.  [your name]